Move util_bloqs.py to bookkeeping/
