+++ qualtran/_infra/gate_with_registers_test.py
@@ -30,7 +30,7 @@ from qualtran import (
     SoquetT,
 )
 from qualtran.bloqs.basic_gates import XGate, YGate, ZGate
-from qualtran.bloqs.util_bloqs import Power
+from qualtran.bloqs.bookkeeping import Power
 from qualtran.testing import execute_notebook
 
 if TYPE_CHECKING:
