+++ qualtran/bloqs/bookkeeping/util_bloqs_test.py
@@ -25,7 +25,15 @@ from qualtran._infra.gate_with_registers import GateWithRegisters, get_named_qub
 from qualtran.bloqs.basic_gates import CNOT, XGate
 from qualtran.bloqs.for_testing import TestAtom, TestCastToFrom, TestMultiRegister
 from qualtran.bloqs.for_testing.atom import TestGWRAtom
-from qualtran.bloqs.util_bloqs import Allocate, Cast, Free, Join, Partition, Power, Split
+from qualtran.bloqs.bookkeeping import (
+    Allocate,
+    Cast,
+    Free,
+    Join,
+    Partition,
+    Power,
+    Split,
+)
 from qualtran.simulation.classical_sim import call_cbloq_classically
 from qualtran.simulation.tensor import bloq_to_dense, cbloq_to_quimb
 from qualtran.testing import assert_valid_bloq_decomposition, execute_notebook
