+++ qualtran/bloqs/chemistry/thc/notebook_utils.py
@@ -19,7 +19,7 @@ import cirq
 from qualtran import Bloq
 from qualtran.bloqs.data_loading.qrom import QROM
 from qualtran.bloqs.data_loading.select_swap_qrom import SelectSwapQROM
-from qualtran.bloqs.util_bloqs import ArbitraryClifford
+from qualtran.bloqs.bookkeeping import ArbitraryClifford
 from qualtran.cirq_interop import CirqGateAsBloq
 from qualtran.resource_counting import SympySymbolAllocator
 from qualtran.resource_counting.generalizers import (
