+++ qualtran/bloqs/factoring/mod_exp_test.py
@@ -22,7 +22,7 @@ import sympy
 from qualtran import Bloq
 from qualtran.bloqs.factoring.mod_exp import _modexp, _modexp_symb, ModExp
 from qualtran.bloqs.factoring.mod_mul import CtrlModMul
-from qualtran.bloqs.util_bloqs import Join, Split
+from qualtran.bloqs.bookkeeping import Join, Split
 from qualtran.drawing import Text
 from qualtran.resource_counting import SympySymbolAllocator
 from qualtran.testing import execute_notebook
