+++ qualtran/bloqs/swap_network/cswap_approx_test.py
@@ -27,7 +27,7 @@ from qualtran.bloqs.swap_network.cswap_approx import (
     _approx_cswap_symb,
     CSwapApprox,
 )
-from qualtran.bloqs.util_bloqs import ArbitraryClifford
+from qualtran.bloqs.bookkeeping import ArbitraryClifford
 from qualtran.cirq_interop.t_complexity_protocol import t_complexity, TComplexity
 from qualtran.testing import assert_valid_bloq_decomposition, execute_notebook
