+++ qualtran/cirq_interop/_cirq_to_bloq.py
@@ -271,7 +271,7 @@ def _ensure_in_reg_exists(
     bb: BloqBuilder, in_reg: _QReg, qreg_to_qvar: Dict[_QReg, Soquet]
 ) -> None:
     """Takes care of qubit allocations, split and joins to ensure `qreg_to_qvar[in_reg]` exists."""
-    from qualtran.bloqs.util_bloqs import Cast
+    from qualtran.bloqs.bookkeeping import Cast
 
     all_mapped_qubits = {q for qreg in qreg_to_qvar for q in qreg.qubits}
     qubits_to_allocate: List[cirq.Qid] = [q for q in in_reg.qubits if q not in all_mapped_qubits]
